Initial commit
Includes SQL file (by Darshan with edits by Ren) to create the database and input dummy data into the tables. Need to double check table structure to see if it correctly matches the diagram we are working with or needs improvement. Also includes backend files (By Mona with edits by Ren) which currently is able to correctly connect to and run a query in the database.